backend/pyproject.toml
[project]
name = "gestor-dividas-pessoais-backend"
version = "0.1.0"
description = "API para gestão de dívidas pessoais"
requires-python = ">=3.8"
dependencies = [
    "annotated-types==0.7.0",
    "anyio>=4.8.0",
    "click==8.1.8",
    "colorama==0.4.6",
    "fastapi==0.115.6",
    "h11==0.14.0",
    "idna==3.10",
    "mysql-connector-python==9.1.0",
    "pydantic==2.10.4",
    "pydantic-core==2.27.2",
    "pydantic-settings==2.7.1",
    "python-dotenv==1.0.1",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.36",
    "starlette==0.41.3",
    "typing-extensions==4.12.2",
    "uvicorn==0.34.0",
    "email-validator",
    "python-jose[cryptography]",
    "python-multipart",
    "bcrypt==4.0.1",
    "passlib[bcrypt]==1.7.4",
    "exceptiongroup>=1.1.3",
    "alembic==1.13.1",
    "python-json-logger==2.0.7",
    "psutil==5.9.8",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "flake8",
    "mypy",
    "httpx"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = """
    --cov=src 
    --cov-report=term-missing 
    --cov-report=html 
    --cov-fail-under=80
    --verbosity=2
    --showlocals
    --tb=short
"""
